feat(playback): clicking a track honors fade on skip
M7 wired "Fade on Skip" into manual Next/Previous only; every click that
starts a track — queue click / play-from-here (play_entry_from_queue,
play_song_from_queue) and browse-view album/playlist/song plays
(play_songs_from_index) — still hard-cut. All three funnel through
PlaybackController::load_play_and_set_current, which now routes via the
module-level plan_click_play:

- Crossfade mode with the engine audibly playing a finite stream (and not
  bit-perfect Strict): reuse the M7 three-phase machinery — stamp the
  shared skip_fade_seq, discharge the click's NextTrackResetEffect and run
  plan_skip_fade under the engine lock (cancel prepared/live transition,
  bump_for_user_action, latch the generation-keyed pending window), build
  the incoming decoder lock-free via the registry path, then fire
  crossfade_to_next with all its re-applied gates. Blocked/failed blends
  fall back to the boundary ease-out + hard load; a competing click or
  skip during the build wins via the seq + generation snapshots.
- Boundary Fade mode: the M7 run_skip_out_fade ease-out (self-refusing)
  before the normal hard load; M2's onset ramp covers the incoming.
- Mode Off, paused/stopped, a radio outgoing (M6's switch-fade domain),
  or bit-perfect Strict: today's hard path, byte-identical — Strict is
  pre-gated by the new click_skip_crossfade_viable so a click never pays
  a wasted decoder build for a blend the format gate would refuse.

Adds TransitionReason::Click for the plan's log line, extends the
"Fade on Skip" settings copy and CHANGELOG bullet to name clicking a
track, and notes the click routing in the audio-engine rules.

+++ data/src/services/settings.rs
@@ -2631,7 +2631,7 @@ name = "sentinel preset"
         assert_eq!(out.font_family, src.font_family, "font_family residual");
         // 2. artwork_column_width_pct (f32, Interface).
         assert!(
-            (out.artwork_column_width_pct - src.artwork_column_width_pct as f32).abs() < 1e-4,
+            (out.artwork_column_width_pct - src.artwork_column_width_pct).abs() < 1e-4,
             "artwork_column_width_pct residual: {} vs {}",
             out.artwork_column_width_pct,
             src.artwork_column_width_pct
